# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Sanko Read API
  description: |
    前端阅读应用（sanko_read）与后端的接口契约。

    **说明**
    - 本文档由前端 `sanko_read/src/api/` 整理，作为后端实现依据
    - 路径与字段名请保持一致；若有调整需前后端协商后同步更新本文档
    - 后端实现后可使用 Knife4j（`/doc.html`）维护在线文档
    - 本文档可导入 Apifox / Postman 进行 Mock 与联调

    **前端环境变量**
    - `VITE_API_BASE_URL`：默认 `http://localhost:8083`
    - `VITE_USE_MOCK=false` 时走真实后端

    **完整需求文档**
    - 详见 [`BACKEND_API.md`](./BACKEND_API.md)
  version: 1.1.0
  contact:
    name: Sanko Read Frontend
servers:
  - url: http://localhost:8083
    description: 本地开发
...
security:
  - BearerAuth: []
paths:
...
  /api/books:
    get:
      tags: [图书]
      summary: 书库列表（不含回收站）
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Book'
    post:
      tags: [图书]
      summary: 将书城书加入个人库
      description: 用户从书城点「开始阅读」或「加入书架」时调用；bookId 应与书城书籍 ID 关联
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Book'
...
  /api/books/{id}:
    get:
      tags: [图书]
      summary: 图书详情
      parameters:
        - $ref: '#/components/parameters/BookId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Book'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [图书]
      summary: 删除图书
      description: |
        - 无 body 或 `permanent` 为 false：从书库删除（是否进回收站由服务端结合用户设置 `disableRecycleBin` 决定）
        - `permanent: true`：从回收站永久删除
      parameters:
        - $ref: '#/components/parameters/BookId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteBookRequest'
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: 'null'
...
  /api/bookshelves:
    get:
      tags: [书架]
      summary: 书架列表
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Bookshelf'
    post:
      tags: [书架]
      summary: 创建书架
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookshelfRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Bookshelf'
  /api/bookshelves/reorder:
    put:
      tags: [书架]
      summary: 书架排序
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReorderBookshelvesRequest'
      responses:
        '200':
          description: 成功，返回排序后的完整列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Bookshelf'
  /api/bookshelves/{id}:
    patch:
      tags: [书架]
      summary: 重命名书架
      parameters:
        - $ref: '#/components/parameters/ShelfId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RenameBookshelfRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Bookshelf'
    delete:
      tags: [书架]
      summary: 删除书架
      parameters:
        - $ref: '#/components/parameters/ShelfId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: 'null'
  /api/bookshelves/{id}/books:
    post:
      tags: [书架]
      summary: 添加图书到书架
      parameters:
        - $ref: '#/components/parameters/ShelfId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddBookToShelfRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: 'null'
  /api/bookshelves/{id}/books/{bookId}:
    delete:
      tags: [书架]
      summary: 从书架移除图书
      parameters:
        - $ref: '#/components/parameters/ShelfId'
        - $ref: '#/components/parameters/BookIdPath'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: 'null'
...
  /api/catalog/books:
    get:
      tags: [书城]
      summary: 书城书籍列表（分类筛选）
      security: []
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
          description: 连载状态，如 serial / finished
        - name: attr
          in: query
          schema:
            type: string
          description: 属性，如 free / vip
        - name: tags
          in: query
          schema:
            type: string
          description: 标签，可多选（逗号分隔）
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            default: 20
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CatalogBookPage'
  /api/catalog/books/{id}:
    get:
      tags: [书城]
      summary: 书城书籍详情
      security: []
      parameters:
        - $ref: '#/components/parameters/CatalogBookId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CatalogBook'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "登录后获得的 token，前端请求头 Authorization: Bearer {token}"
  parameters:
    BookId:
      name: id
      in: path
      required: true
      schema:
        type: string
    BookIdPath:
      name: bookId
      in: path
      required: true
      schema:
        type: string
    ShelfId:
      name: id
      in: path
      required: true
      schema:
        type: string
    CatalogBookId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: 书城书籍 ID
...
  responses:
...
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
  schemas:
    ApiResponse:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
          description: 0 表示成功，非 0 表示业务错误
          example: 0
        message:
          type: string
          example: ok
        data:
          nullable: true
          description: 业务数据，类型随接口变化
...
    Book:
      type: object
      required:
        - id
        - title
        - author
        - progress
        - coverColor
        - coverTitle
        - fileSize
        - format
        - addedAt
        - category
      properties:
        id:
          type: string
        title:
          type: string
        author:
          type: string
        progress:
          type: number
          format: float
          description: 阅读进度百分比，如 36.0
        coverColor:
          type: string
          description: 封面背景色，如 "#e86c1a"
        coverTitle:
          type: string
        coverSubtitle:
          type: string
        coverUrl:
          type: string
          description: 封面图片 URL（可选）
        fileSize:
          type: string
          description: 人类可读大小，如 "12.6 MB"
        format:
          type: string
          enum: [PDF, EPUB]
        addedAt:
          type: string
          description: 添加日期，格式 YYYY-MM-DD
        category:
          type: string
        lastReadAt:
          type: string
          description: 最后阅读日期，格式 YYYY-MM-DD
...
    DeleteBookRequest:
      type: object
      properties:
        permanent:
          type: boolean
          description: true 表示永久删除（通常用于回收站）
...
    Bookshelf:
      type: object
      required: [id, name, bookIds]
      properties:
        id:
          type: string
        name:
          type: string
        bookIds:
          type: array
          items:
            type: string
    CreateBookshelfRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
    RenameBookshelfRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
    ReorderBookshelvesRequest:
      type: object
      required: [orderedIds]
      properties:
        orderedIds:
          type: array
          items:
            type: string
    AddBookToShelfRequest:
      type: object
      required: [bookId]
      properties:
        bookId:
          type: string
...
    CatalogBook:
      type: object
      required: [id, title, author, coverColor, coverTitle]
      properties:
        id:
          type: string
        title:
          type: string
        author:
          type: string
        coverColor:
          type: string
        coverTitle:
          type: string
        coverUrl:
          type: string
        description:
          type: string
        category:
          type: string
        updateStatus:
          type: string
        latestChapter:
          type: string
        tags:
          type: array
          items:
            type: string
        synopsis:
          type: string
...
    CatalogBookPage:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CatalogBook'
        total:
          type: integer
...